requirements.txt
python-docx==1.2.0
python-dotenv==1.0.1
openai==2.7.1
cohere==5.20.0
sentence-transformers==5.1.2
faiss-cpu==1.12.0
numpy==2.3.4
tqdm==4.67.1
pandas==2.3.3
ragas==0.3.8
datasets==4.4.1
python-toon==0.1.3